requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
